+++ src/org/neuroph/core/input/EuclideanRBF.java
@@ -4,9 +4,7 @@
 package org.neuroph.core.input;
 
 import java.util.List;
-import org.neuroph.core.Connection;
-import org.neuroph.core.Neuron;
-import org.neuroph.core.Weight;
+import org.neuroph.core.*;
 
 /**
  * @author zoran
@@ -25,7 +23,7 @@ public class EuclideanRBF extends InputFunction {
 			sqrSum += diff * diff;
 		}
 
-		return 0.5 * Math.sqrt(sqrSum) / (double) inputConnections.size(); // ovo trebaprebaciti u novu funkciju transfera sa odgovarajucim izvodom
+		return 0.5 * Math.sqrt(sqrSum) / inputConnections.size(); // ovo trebaprebaciti u novu funkciju transfera sa odgovarajucim izvodom
 	}
 
 }
